fix: 🐛 change playground tour
